app/config/requirements.txt
asyncpg
aioredis
aiogram
python-dotenv
gino